requirements.txt
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
idna==2.6
numpy==1.13.3
opencv-python==3.3.0.10
pycparser==2.18
python-librtmp==0.3.0
requests==2.18.4
urllib3==1.22
